修改删除节点bug 去年ehcache缓存，用的是spring-boot自带的缓存，其实只是去掉了ehcache的jar和配置文件
另外还增加了节点话题数的校正功能

+++ src/main/java/co/yiiu/web/admin/NodeAdminController.java
@@ -3,6 +3,7 @@ package co.yiiu.web.admin;
 import co.yiiu.core.base.BaseController;
 import co.yiiu.module.node.model.Node;
 import co.yiiu.module.node.service.NodeService;
+import co.yiiu.module.topic.service.TopicService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletResponse;
 import java.util.Date;
+import java.util.List;
 
 /**
  * Created by tomoya.
@@ -25,11 +27,13 @@ public class NodeAdminController extends BaseController {
 
   @Autowired
   private NodeService nodeService;
+  @Autowired
+  private TopicService topicService;
 
   @GetMapping("/list")
   public String list(Model model, Integer pid) {
     model.addAttribute("pnodes", nodeService.findByPid(0));
-    if(pid == null || pid == 0) {
+    if (pid == null || pid == 0) {
       model.addAttribute("nodes", nodeService.findAll(true));
     } else {
       model.addAttribute("pid", pid);
@@ -55,7 +59,7 @@ public class NodeAdminController extends BaseController {
     node.setIntro(intro);
     node.setInTime(new Date());
     nodeService.save(node);
-    if(pid == 0) pid = node.getId();
+    if (pid == 0) pid = node.getId();
     return redirect("/admin/node/list?pid=" + pid);
   }
 
@@ -79,10 +83,42 @@ public class NodeAdminController extends BaseController {
   }
 
   @GetMapping("/{id}/delete")
-  public String delete(@PathVariable int id, HttpServletResponse response) {
+  public String delete(@PathVariable int id, HttpServletResponse response) throws Exception {
+    nodeService.clearCache();// 缓存导致展示列表后，再用id查询node信息的时候会返回一个集合，所以在查询之前先清一下node的缓存
     Node node = nodeService.findById(id);
+    if (node.getTopicCount() > 0) throw new Exception("这个节点下还有话题，不能删除");
+
     int pid = node.getPid();
     nodeService.deleteById(id);
-    return redirect(response, "/admin/node/list?pid=" + pid);
+    if (pid > 0) {
+      return redirect(response, "/admin/node/list?pid=" + pid);
+    } else {
+      nodeService.deleteByPid(id);
+      return redirect(response, "/admin/node/list");
+    }
+  }
+
+  /**
+   * 对节点下的话题数进行校正
+   * @param response
+   * @return
+   */
+  @GetMapping("/correction")
+  public String correction(HttpServletResponse response) {
+    nodeService.clearCache();
+    List<Node> pNodes = nodeService.findByPid(0);
+    pNodes.forEach(pNode -> {
+      pNode.setTopicCount(0);
+      List<Node> cNodes = nodeService.findByPid(pNode.getId());
+      cNodes.forEach(cNode -> {
+        long topicCount = topicService.countByNode(cNode);
+        cNode.setTopicCount((int) topicCount);
+        nodeService.save(cNode);
+
+        pNode.setTopicCount(pNode.getTopicCount() + (int) topicCount);
+      });
+      nodeService.save(pNode);
+    });
+    return redirect(response, "/admin/node/list");
   }
 }
